requirements.txt
Flask==2.1.1
Flask-Cors==3.1.1
tensorflow==2.10.0
pillow==9.2.0
numpy==1.23.2
scikit-learn==1.1.1
matplotlib==3.5.2
gunicorn==20.1.0
requests==2.28.1
expo==46.0.0
react==18.0.0
react-native==0.68.0
expo-image-picker==12.0.1
axios==0.27.2
react-navigation==4.4.4
react-navigation-stack==2.10.4
